+++ src/data/types.ts
@@ -15,3 +15,4 @@ export type RoleRow = Tables['roles']['Row'];
 export type RoleID = RoleRow['id'];
 export type VolunteerRow = Tables['volunteers']['Row'];
 export type VolunteerID = VolunteerRow['id'];
+export type Response = Database['public']['Enums']['invited'];